Add custom format checker for array elements; new format_checkers module
