+++ simtools/schemas/model_parameters/array_layouts.schema.yml
@@ -21,6 +21,7 @@ data:
             type: array
             items:
               type: string
+              format: array_element
         required:
           - name
           - elements
